+++ projects/multi-entries-app/src/Nav.tsx
@@ -3,6 +3,7 @@ const Nav = () => (
   <div className="nav">
     <a href="/">home</a>
     <a href="/info.html">info</a>
+    <a href="/info-pc.html">pc info</a>
     <a href="/work/index.html">work</a>
   </div>
 )