+++ projects/multi-entries-app/src/components/pc.tsx
@@ -1,0 +1,10 @@
+import Nav from 'src/Nav'
+import StoreComp from 'src/StoreComp'
+const PCInfo = () => (
+  <>
+    <Nav />
+    <h1>PCInfo</h1>
+    <StoreComp />
+  </>
+)
+export default PCInfo