+++ projects/multi-entries-app/src/pc/index.tsx
@@ -1,0 +1,3 @@
+import ReactDOM from 'react-dom'
+import PCInfo from 'src/components/pc'
+ReactDOM.render(<PCInfo />, document.getElementById('emp-root'))